+++ more_abc/__init__.py
@@ -1,0 +1,44 @@
+# Copyright (C) 2026 Evan Yang <quantbit@126.com>
+
+# MIT License
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+# This module refers to PEP 3119.
+# See https://peps.python.org/pep-3119/ or https://legacy.python.org/dev/peps/pep-3119/
+"""
+This module is an extension of the `abc` module, 
+with many similar features added.
+"""
+
+from .more import (ABCMixin, 
+                   ABCclassType, 
+                   ABCMetaclassType, 
+                   ABCException, 
+                   ABCWarning)
+
+__all__ = ["ABCMixin", 
+           "ABCclassType", 
+           "ABCMetaclassType",
+           "ABCException",
+           "ABCWarning",
+           "version"]
+
+__version__ = "0.0.4"
+version = __version__